+++ Sources/VComponents/Components/Buttons/Plain Button/VPlainButtonUIModel.swift
@@ -33,21 +33,8 @@ public struct VPlainButtonUIModel {
     /// Model that contains layout properties.
     public struct Layout {
         // MARK: Properties
-        /// Icon size.
-        /// Set to `16x16` on `iOS`.
-        /// Set to `16x16` on `macOS`.
-        /// Set to `18x18` on `watchOS`
-        public var iconSize: CGSize = {
-#if os(iOS)
-            return CGSize(dimension: 16)
-#elseif os(macOS)
-            return CGSize(dimension: 16)
-#elseif os(watchOS)
-            return CGSize(dimension: 18)
-#else
-            fatalError() // Not supported
-#endif
-        }()
+        /// Icon size. Set to `18x18`.
+        public var iconSize: CGSize = .init(dimension: 18)
         
         /// Title text minimum scale factor. Set to `0.75`.
         public var titleTextMinimumScaleFactor: CGFloat = GlobalUIModel.Common.minimumScaleFactor